+++ Lock/SingleInputView.swift
@@ -73,6 +73,19 @@ class SingleInputView: UIView, Form, Stylable {
         }
     }
 
+    var onSubmit: (InputField) -> Bool {
+        get {
+            return self.inputField.onSubmit
+        }
+        set {
+            self.inputField.onSubmit = newValue
+        }
+    }
+
+    func shouldSubmit() -> Bool {
+        return self.inputField.onSubmit(self.inputField)
+    }
+
     func needsToUpdateState() {
         self.inputField.needsToUpdateState()
     }