+++ LockTests/Presenters/DatabasePresenterSpec.swift
@@ -86,6 +86,20 @@ class DatabasePresenterSpec: QuickSpec {
             }
         }
 
+        describe("form submit") {
+
+            it("should allow form submit") {
+                view.form?.onSubmit = { _ in return true }
+                expect(view.form?.shouldSubmit()) == true
+            }
+
+            it("should not allow form submit") {
+                view.form?.onSubmit = { _ in return false }
+                expect(view.form?.shouldSubmit()) == false
+            }
+
+        }
+
         describe("accept terms alert") {
 
             it("present alert") {
@@ -263,10 +277,30 @@ class DatabasePresenterSpec: QuickSpec {
                     expect(interactor.email) == email
                 }
 
+                it("should allow a valid email") {
+                    let input = mockInput(.email, value: email)
+                    expect(view.form?.onSubmit(input)) == true
+                }
+
+                it("should not allow an invalid email") {
+                    let input = mockInput(.email, value: "invalid")
+                    expect(view.form?.onSubmit(input)) == false
+                }
+
                 it("should update username") {
                     let input = mockInput(.username, value: username)
                     view.form?.onValueChange(input)
                     expect(interactor.username) == username
+                }
+
+                it("should allow a valid username") {
+                    let input = mockInput(.username, value: username)
+                    expect(view.form?.onSubmit(input)) == true
+                }
+
+                it("should not allow an invalid username") {
+                    let input = mockInput(.username, value: "invalid")
+                    expect(view.form?.onSubmit(input)) == false
                 }
 
                 it("should update password") {
@@ -275,10 +309,30 @@ class DatabasePresenterSpec: QuickSpec {
                     expect(interactor.password) == password
                 }
 
+                it("should allow a valid password") {
+                    let input = mockInput(.password, value: password)
+                    expect(view.form?.onSubmit(input)) == true
+                }
+
+                it("should not allow an invalid password") {
+                    let input = mockInput(.password, value: "invalid")
+                    expect(view.form?.onSubmit(input)) == false
+                }
+
                 it("should update username or email") {
                     let input = mockInput(.emailOrUsername, value: username)
                     view.form?.onValueChange(input)
                     expect(interactor.username) == username
+                }
+
+                it("should allow a valid username or email") {
+                    let input = mockInput(.emailOrUsername, value: username)
+                    expect(view.form?.onSubmit(input)) == true
+                }
+
+                it("should not allow an invalid username or email") {
+                    let input = mockInput(.emailOrUsername, value: "invalid")
+                    expect(view.form?.onSubmit(input)) == false
                 }
 
                 it("should not update if type is not valid for db connection") {
@@ -466,16 +520,46 @@ class DatabasePresenterSpec: QuickSpec {
                     expect(interactor.email) == email
                 }
 
+                it("should allow a valid email") {
+                    let input = mockInput(.email, value: email)
+                    expect(view.form?.onSubmit(input)) == true
+                }
+
+                it("should not allow an invalid email") {
+                    let input = mockInput(.email, value: "invalid")
+                    expect(view.form?.onSubmit(input)) == false
+                }
+
                 it("should update username") {
                     let input = mockInput(.username, value: username)
                     view.form?.onValueChange(input)
                     expect(interactor.username) == username
                 }
 
+                it("should allow a valid username") {
+                    let input = mockInput(.username, value: username)
+                    expect(view.form?.onSubmit(input)) == true
+                }
+
+                it("should not allow an invalid username") {
+                    let input = mockInput(.username, value: "invalid")
+                    expect(view.form?.onSubmit(input)) == false
+                }
+
                 it("should update password") {
                     let input = mockInput(.password, value: password)
                     view.form?.onValueChange(input)
                     expect(interactor.password) == password
+                }
+
+                it("should allow a valid password") {
+                    let input = mockInput(.password, value: password)
+                    expect(view.form?.onSubmit(input)) == true
+                }
+
+                it("should not allow an invalid password") {
+                    let input = mockInput(.password, value: "invalid")
+                    expect(view.form?.onSubmit(input)) == false
                 }
 
                 it("should custom field") {
@@ -839,6 +923,7 @@ class DatabasePresenterSpec: QuickSpec {
                         presenter.enterpriseInteractor = enterpriseInteractor
                         
                         view = presenter.view as? DatabaseOnlyView
+                        view.form?.onSubmit = { _ in return true }
                         
                         let input = mockInput(.email, value: "user@valid.com")
                         view.form?.onValueChange(input)